% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncucyteFlow-package.R
\docType{package}
\name{IncucyteFlow-package}
\alias{IncucyteFlow}
\alias{IncucyteFlow-package}
\title{IncucyteFlow: Convert Incucyte Microscopy Data to FCS Format for Flow Cytometry Analysis}
\description{
Provides tools to parse Incucyte platemap XML files, read and tidy Incucyte object-level CSV data, convert to FCS files compatible with flow cytometry software, and load into GatingSet objects for downstream analysis with flowWorkspace and CytoExploreR.
}
\author{
\strong{Maintainer}: Brian Reismann \email{breisma1@jh.edu.com}

}
\keyword{internal}
